+++ R/Archive.R
@@ -4,10 +4,17 @@
 #' Container around a [data.table::data.table] which stores all performed
 #' function calls of the Objective.
 #'
+#' @section S3 Methods:
+#' * `as.data.table(archive)`\cr
+#'   [Archive] -> [data.table::data.table()]\cr
+#'   Returns a tabular view of all performed function calls of the
+#'   Objective. The `x_domain` column is unnested to separate columns.
+#'
 #' @template param_codomain
 #' @template param_search_space
 #' @template param_xdt
 #' @template param_ydt
+#' @template param_store_x_domain
 #' @export
 Archive = R6Class("Archive",
   public = list(
@@ -20,15 +27,23 @@ Archive = R6Class("Archive",
     #' Codomain of objective function.
     codomain = NULL,
 
-    #' @field start_time ([POSIXct]).
+    #' @field start_time ([POSIXct])\cr
+    #' Time stamp of when the optimization started. The time is set by the
+    #' [Optimizer].
     start_time = NULL,
 
-    #' @field check_values (`logical(1)`)
+    #' @field check_values (`logical(1)`)\cr
+    #' Determines if points and results are checked for validity.
     check_values = NULL,
 
     #' @field data ([data.table::data.table])\cr
     #' Contains all performed [Objective] function calls.
     data = NULL,
+
+    #' @field store_x_domain (`logical(1)`)\cr
+    #' Determines if x values, should be stored in `$data$x_domain` as list
+    #' items. The trafo will be applied if defined in `search_space`.
+    store_x_domain = NULL,
 
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
@@ -36,11 +51,12 @@ Archive = R6Class("Archive",
     #' @param check_values (`logical(1)`)\cr
     #' Should x-values that are added to the archive be checked for validity?
     #' Search space that is logged into archive.
-    initialize = function(search_space, codomain, check_values = TRUE) {
+    initialize = function(search_space, codomain, check_values = TRUE, store_x_domain = TRUE) {
       self$search_space = assert_param_set(search_space)
       self$codomain = assert_param_set(codomain)
       self$check_values = assert_flag(check_values)
       self$data = data.table()
+      self$store_x_domain = assert_flag(store_x_domain)
     },
 
     #' @description
@@ -48,10 +64,11 @@ Archive = R6Class("Archive",
     #'
     #' @param xss_trafoed (`list()`)\cr
     #' Transformed point(s) in the *domain space*.
-    add_evals = function(xdt, xss_trafoed, ydt) {
+    #' Not stored and needed if `store_x_domain = FALSE`.
+    add_evals = function(xdt, xss_trafoed = NULL, ydt) {
       assert_data_table(xdt)
       assert_data_table(ydt)
-      assert_list(xss_trafoed)
+      assert_list(xss_trafoed, null.ok = !self$store_x_domain)
       assert_data_table(ydt[, self$cols_y, with = FALSE], any.missing = FALSE)
       if (self$check_values) {
         self$search_space$assert_dt(xdt[, self$cols_x, with = FALSE])
@@ -59,7 +76,9 @@ Archive = R6Class("Archive",
       xydt = cbind(xdt, ydt)
       assert_subset(c(self$search_space$ids(), self$codomain$ids()), colnames(xydt))
       batch_nr = self$data$batch_nr
-      set(xydt, j = "x_domain", value = list(xss_trafoed))
+      if (self$store_x_domain) {
+        set(xydt, j = "x_domain", value = list(xss_trafoed))
+      }
       set(xydt, j = "timestamp", value = Sys.time())
       set(xydt, j = "batch_nr", value = if (length(batch_nr)) max(batch_nr) + 1L else 1L)
       self$data = rbindlist(list(self$data, xydt), fill = TRUE, use.names = TRUE)
@@ -139,11 +158,11 @@ Archive = R6Class("Archive",
       }
     },
 
-    #' @field cols_x (`character()`).
+    #' @field cols_x (`character()`)\cr
     #' Column names of search space parameters.
     cols_x = function() self$search_space$ids(),
 
-    #' @field cols_y (`character()`).
+    #' @field cols_y (`character()`)\cr
     #' Column names of codomain parameters.
     cols_y = function() self$codomain$ids()
   ),
@@ -155,7 +174,7 @@ Archive = R6Class("Archive",
 
 #' @export
 as.data.table.Archive = function(x, ...) { # nolint
-  if (nrow(x$data)==0) {
+  if (!x$store_x_domain || nrow(x$data)==0) {
     copy(x$data)
   } else {
     unnest(copy(x$data), "x_domain", prefix = "{col}_")