+++ R/OptimizerCmaes.R
@@ -12,18 +12,22 @@
 #'
 #' @section Parameters:
 #' \describe{
-#' \item{`par`}{`numeric()`}
 #' \item{`sigma`}{`numeric(1)`}
+#' \item{`start_values`}{`character(1)`\cr
+#' Create `random` start values or based on `center` of search space? In the 
+#' latter case, it is the center of the parameters before a trafo is applied.}
 #' }
 #'
 #' For the meaning of the control parameters, see [adagio::pureCMAES()]. Note
 #' that we have removed all control parameters which refer to the termination of
 #' the algorithm and where our terminators allow to obtain the same behavior.
+#' 
+#' @template section_progress_bars
 #'
 #' @export
 #' @examples
+#' if(requireNamespace("adagio")) {
 #' library(paradox)
-#' library(data.table)
 #'
 #' domain = ParamSet$new(list(ParamDbl$new("x", lower = -1, upper = 1)))
 #'
@@ -38,12 +42,14 @@
 #' objective = ObjectiveRFun$new(fun = objective_function,
 #'                               domain = domain,
 #'                               codomain = codomain)
-#' terminator = trm("evals", n_evals = 2)
-#' instance = OptimInstanceSingleCrit$new(objective = objective,
-#'                              search_space = search_space,
-#'                              terminator = terminator)
+#' terminator = trm("evals", n_evals = 10)
+#' instance = OptimInstanceSingleCrit$new(
+#'  objective = objective,
+#'  search_space = search_space,
+#'  terminator = terminator)
 #'
-#' optimizer = opt("cmaes", par = 1)
+#'
+#' optimizer = opt("cmaes")
 #'
 #' # Modifies the instance by reference
 #' optimizer$optimize(instance)
@@ -52,7 +58,8 @@
 #' instance$result
 #'
 #' # Allows access of data.table of full path of all evaluations
-#' as.data.table(instance$archive)
+#' as.data.table(instance$archive$data)
+#' }
 OptimizerCmaes = R6Class("OptimizerCmaes",
   inherit = Optimizer,
   public = list(
@@ -61,9 +68,10 @@ OptimizerCmaes = R6Class("OptimizerCmaes",
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
       ps = ParamSet$new(list(
-        ParamUty$new("par", tags = "required"),
-        ParamDbl$new("sigma", default = 0.5)
+        ParamDbl$new("sigma", default = 0.5),
+        ParamFct$new("start_values", default = "random", levels = c("random", "center"))
       ))
+      ps$values$start_values = "random"
       super$initialize(
         param_set = ps,
         param_classes = "ParamDbl",
@@ -76,6 +84,8 @@ OptimizerCmaes = R6Class("OptimizerCmaes",
   private = list(
     .optimize = function(inst) {
       pv = self$param_set$values
+      pv$par = search_start(inst$search_space, type = pv$start_values)
+      pv$start_values = NULL
       pv$stopeval = .Machine$integer.max # make sure pureCMAES does not stop
       pv$stopfitness = -Inf