+++ R/OptimizerGenSA.R
@@ -22,12 +22,47 @@
 #' For the meaning of the control parameters, see [GenSA::GenSA()]. Note that we
 #' have removed all control parameters which refer to the termination of the
 #' algorithm and where our terminators allow to obtain the same behavior.
+#' 
+#' @template section_progress_bars
 #'
 #' @source
 #' `r format_bib("tsallis_1996", "xiang_2013")`
 #'
 #' @export
-#' @template example
+#' @examples
+#' if(requireNamespace("GenSA")) {
+#' library(paradox)
+#'
+#' domain = ParamSet$new(list(ParamDbl$new("x", lower = -1, upper = 1)))
+#'
+#' search_space = ParamSet$new(list(ParamDbl$new("x", lower = -1, upper = 1)))
+#'
+#' codomain = ParamSet$new(list(ParamDbl$new("y", tags = "minimize")))
+#'
+#' objective_function = function(xs) {
+#'   list(y = as.numeric(xs)^2)
+#' }
+#'
+#' objective = ObjectiveRFun$new(fun = objective_function,
+#'                               domain = domain,
+#'                               codomain = codomain)
+#' terminator = trm("evals", n_evals = 10)
+#' instance = OptimInstanceSingleCrit$new(
+#'  objective = objective,
+#'  search_space = search_space,
+#'  terminator = terminator)
+#'
+#' optimizer = opt("cmaes")
+#'
+#' # Modifies the instance by reference
+#' optimizer$optimize(instance)
+#'
+#' # Returns best scoring evaluation
+#' instance$result
+#'
+#' # Allows access of data.table of full path of all evaluations
+#' as.data.table(instance$archive$data)
+#' }
 OptimizerGenSA = R6Class("OptimizerGenSA", inherit = Optimizer,
   public = list(