+++ R/OptimizerNLoptr.R
@@ -10,12 +10,14 @@
 #' @section Parameters:
 #' \describe{
 #' \item{`algorithm`}{`character(1)`}
-#' \item{`x0`}{`numeric()`}
 #' \item{`eval_g_ineq`}{`function()`}
 #' \item{`xtol_rel`}{`numeric(1)`}
 #' \item{`xtol_abs`}{`numeric(1)`}
 #' \item{`ftol_rel`}{`numeric(1)`}
 #' \item{`ftol_abs`}{`numeric(1)`}
+#' \item{`start_values`}{`character(1)`\cr
+#' Create `random` start values or based on `center` of search space? In the 
+#' latter case, it is the center of the parameters before a trafo is applied.}
 #' }
 #'
 #' For the meaning of the control parameters, see [nloptr::nloptr()] and
@@ -27,6 +29,8 @@
 #' (`xtol_rel = 10^-4`, `xtol_abs = rep(0.0, length(x0))`, `ftol_rel = 0.0` and
 #' `ftol_abs = 0.0`) are still available and implemented with their package
 #' defaults. To deactivate these conditions, set them to `-1`.
+#' 
+#' @template section_progress_bars
 #'
 #' @source
 #' `r format_bib("johnson_2014")`
@@ -34,8 +38,8 @@
 #' @export
 #' @examples
 #' \donttest{
+#' if(requireNamespace("nloptr")) {
 #' library(paradox)
-#' library(data.table)
 #'
 #' domain = ParamSet$new(list(ParamDbl$new("x", lower = -1, upper = 1)))
 #'
@@ -53,11 +57,13 @@
 #'
 #' # We use the internal termination criterion xtol_rel
 #' terminator = trm("none")
-#' instance = OptimInstanceSingleCrit$new(objective = objective,
-#'   search_space = search_space,
-#'   terminator = terminator)
+#' instance = OptimInstanceSingleCrit$new(
+#'  objective = objective,
+#'  search_space = search_space,
+#'  terminator = terminator)
 #'
-#' optimizer = opt("nloptr", x0 = 1, algorithm = "NLOPT_LN_BOBYQA")
+#'
+#' optimizer = opt("nloptr", algorithm = "NLOPT_LN_BOBYQA")
 #'
 #' # Modifies the instance by reference
 #' optimizer$optimize(instance)
@@ -67,6 +73,7 @@
 #'
 #' # Allows access of data.table of full path of all evaluations
 #' as.data.table(instance$archive)
+#' }
 #' }
 OptimizerNLoptr = R6Class("OptimizerNLoptr", inherit = Optimizer,
   public = list(
@@ -91,7 +98,6 @@ OptimizerNLoptr = R6Class("OptimizerNLoptr", inherit = Optimizer,
             "NLOPT_LN_AUGLAG", "NLOPT_LD_AUGLAG", "NLOPT_LN_AUGLAG_EQ",
             "NLOPT_LD_AUGLAG_EQ", "NLOPT_LN_BOBYQA", "NLOPT_GN_ISRES"),
         tags = "required"),
-        ParamUty$new("x0", tags = "required"),
         ParamUty$new("eval_g_ineq", default = NULL),
         ParamDbl$new("xtol_rel", default = 10^-4, lower = 0, upper = Inf,
           special_vals = list(-1)),
@@ -100,8 +106,10 @@ OptimizerNLoptr = R6Class("OptimizerNLoptr", inherit = Optimizer,
         ParamDbl$new("ftol_rel", default = 0, lower = 0, upper = Inf,
           special_vals = list(-1)),
         ParamDbl$new("ftol_abs", default = 0, lower = 0, upper = Inf,
-          special_vals = list(-1))
+          special_vals = list(-1)),
+        ParamFct$new("start_values", default = "random", levels = c("random", "center"))
       ))
+      ps$values$start_values = "random"
       super$initialize(param_set = ps, param_classes = "ParamDbl",
         properties = "single-crit", packages = "nloptr")
     }
@@ -110,6 +118,8 @@ OptimizerNLoptr = R6Class("OptimizerNLoptr", inherit = Optimizer,
   private = list(
     .optimize = function(inst) {
       pv = self$param_set$values
+      pv$x0 = search_start(inst$search_space, type = pv$start_values)
+      pv$start_values = NULL
       opts = pv[which(names(pv) %nin% formalArgs(nloptr::nloptr))]
       # Deactivate termination criterions which are replaced by Terminators
       opts = insert_named(opts, list(