+++ tests/testthat/test_TerminatorEvals.R
@@ -1,5 +1,3 @@
-context("TerminatorEvals")
-
 test_that("TerminatorEvals works", {
   inst = MAKE_INST_2D(7L)
   expect_output(print(inst$terminator), "TerminatorEvals")
@@ -24,4 +22,10 @@ test_that("status method works", {
   expect_equal(inst$terminator$status(inst$archive)["max_steps"],  c("max_steps" = 10))
   expect_equal(inst$terminator$status(inst$archive)["current_steps"], c("current_steps"= 2))
   expect_equal(inst$terminator$remaining_time(inst$archive), Inf)
+})
+
+test_that("TerminatorEvals works with empty archive" ,{
+  terminator = TerminatorEvals$new()
+  archive = Archive$new(ps(x = p_dbl()), ps(y = p_dbl()))
+  expect_false(terminator$is_terminated(archive))
 })