+++ tests/testthat/test_TerminatorNone.R
@@ -1,5 +1,3 @@
-context("TerminatorNone")
-
 test_that("TerminatorNone works", {
   terminators = trms(c("evals", "none"))
   terminators[[1]]$param_set$values$n_evals = 10L
@@ -8,4 +6,10 @@ test_that("TerminatorNone works", {
   inst = MAKE_INST_2D(terminator)
   a = random_search(inst, batch_size = 1L)
   expect_equal(a$n_evals, 10L, info = mode)
+})
+
+test_that("TerminatorNone works with empty archive" ,{
+  terminator = TerminatorNone$new()
+  archive = Archive$new(ps(x = p_dbl()), ps(y = p_dbl()))
+  expect_false(terminator$is_terminated(archive))
 })