+++ tests/testthat/test_TerminatorPerfReached.R
@@ -1,5 +1,3 @@
-context("TerminatorPerfReached")
-
 test_that("TerminatorPerfReached works", {
   terminator = TerminatorPerfReached$new()
   terminator$param_set$values$level = 0.2
@@ -13,4 +11,10 @@ test_that("TerminatorPerfReached works", {
 test_that("TerminatorPerfReached in OptimInstanceMultiCrit throws an error", {
   terminator = TerminatorPerfReached$new()
   expect_error(MAKE_INST_2D_2D(terminator))
+})
+
+test_that("TerminatorPerfReached works with empty archive" ,{
+  terminator = TerminatorPerfReached$new()
+  archive = Archive$new(ps(x = p_dbl()), ps(y = p_dbl()))
+  expect_false(terminator$is_terminated(archive))
 })