+++ tests/testthat/test_TerminatorRunTime.R
@@ -1,5 +1,3 @@
-context("TerminatorRunTime")
-
 test_that("TerminatorRunTime works", {
   skip_on_cran()
 
@@ -23,4 +21,10 @@ test_that("max and current works", {
   expect_equal(inst$terminator$status(inst$archive)["max_steps"], c("max_steps" = 3))
   expect_equal(inst$terminator$status(inst$archive)["current_steps"], c("current_steps" = 1), tolerance = 1)
   expect_equal(inst$terminator$remaining_time(inst$archive), 2, tolerance = 1)
+})
+
+test_that("TerminatorRunTime works with empty archive" ,{
+  terminator = TerminatorRunTime$new()
+  archive = Archive$new(ps(x = p_dbl()), ps(y = p_dbl()))
+  expect_false(terminator$is_terminated(archive))
 })