+++ tests/testthat/test_TerminatorStagnationBatch.R
@@ -1,5 +1,3 @@
-context("TerminatorStagnationBatch")
-
 test_that("TerminatorStagnationBatch works with single objective and n = 1", {
   terminator = TerminatorStagnationBatch$new()
   terminator$param_set$values$n = 1
@@ -35,4 +33,10 @@ test_that("TerminatorStagnationBatch works with single objective and n = 2", {
 test_that("TerminatorStagnationBatch in OptimInstanceMultiCrit throws an error", {
   terminator = TerminatorStagnationBatch$new()
   expect_error(MAKE_INST_2D_2D(terminator))
+})
+
+test_that("TerminatorStagnationBatch works with empty archive" ,{
+  terminator = TerminatorStagnationBatch$new()
+  archive = Archive$new(ps(x = p_dbl()), ps(y = p_dbl()))
+  expect_false(terminator$is_terminated(archive))
 })